Update anagrams and fill in the blanks endpoints
Note: Now they do not call the pick_words function anymore. Theme is still passed as a query parameter, but it has no purpose, it's just returned back to the frontend. This should be removed from the list of query parameters in the future and just be passed around in the React component directly from the user selection